putting different objects in files
